+++ tests/integration/taskrouter/v1/workspace/task_queue/test_task_queue_real_time_statistics.py
@@ -33,6 +33,7 @@ class TaskQueueRealTimeStatisticsTestCase(IntegrationTestCase):
             '''
             {
                 "longest_task_waiting_age": 100,
+                "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "task_queue_sid": "WQaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "tasks_by_status": {
                     "reserved": 0,