+++ tests/integration/taskrouter/v1/workspace/test_event.py
@@ -43,10 +43,12 @@ class EventTestCase(IntegrationTestCase):
                     "worker_name": "JustinWorker",
                     "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                     "worker_time_in_previous_activity": "26",
+                    "worker_time_in_previous_activity_ms": "26123",
                     "workspace_name": "WorkspaceName",
                     "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                 },
                 "event_date": "2015-02-07T00:32:41Z",
+                "event_date_ms": 987654321111,
                 "event_type": "worker.activity",
                 "resource_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "resource_type": "worker",
@@ -54,6 +56,7 @@ class EventTestCase(IntegrationTestCase):
                 "sid": "EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "source": "twilio",
                 "source_ip_address": "1.2.3.4",
+                "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
@@ -95,10 +98,12 @@ class EventTestCase(IntegrationTestCase):
                             "worker_name": "JustinWorker",
                             "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                             "worker_time_in_previous_activity": "26",
+                            "worker_time_in_previous_activity_ms": "26123",
                             "workspace_name": "WorkspaceName",
                             "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                         },
                         "event_date": "2015-02-07T00:32:41Z",
+                        "event_date_ms": 987654321111,
                         "event_type": "worker.activity",
                         "resource_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "resource_type": "worker",
@@ -106,6 +111,7 @@ class EventTestCase(IntegrationTestCase):
                         "sid": "EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "source": "twilio",
                         "source_ip_address": "1.2.3.4",
+                        "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Events/EVaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],