+++ tests/integration/taskrouter/v1/workspace/test_workspace_statistics.py
@@ -83,7 +83,7 @@ class WorkspaceStatisticsTestCase(IntegrationTestCase):
                         }
                     ],
                     "longest_task_waiting_age": 0,
-                    "longest_task_waiting_sid": null,
+                    "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                     "tasks_by_status": {
                         "assigned": 0,
                         "pending": 0,