+++ tests/integration/taskrouter/v1/workspace/workflow/test_workflow_real_time_statistics.py
@@ -33,6 +33,7 @@ class WorkflowRealTimeStatisticsTestCase(IntegrationTestCase):
             '''
             {
                 "longest_task_waiting_age": 100,
+                "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workflows/WWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/RealTimeStatistics",
                 "tasks_by_priority": {},
                 "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",