+++ tests/integration/taskrouter/v1/workspace/workflow/test_workflow_statistics.py
@@ -48,7 +48,7 @@ class WorkflowStatisticsTestCase(IntegrationTestCase):
                 },
                 "realtime": {
                     "longest_task_waiting_age": 0,
-                    "longest_task_waiting_sid": null,
+                    "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                     "tasks_by_status": {
                         "assigned": 1,
                         "pending": 0,