+++ tests/integration/verify/v1/service/test_verification.py
@@ -51,7 +51,9 @@ class VerificationTestCase(IntegrationTestCase):
                         "mobile_network_code": "150",
                         "type": "mobile"
                     }
-                }
+                },
+                "amount": "$29.99",
+                "payee": "Acme"
             }
             '''
         ))