+++ tests/integration/verify/v1/service/test_verification_check.py
@@ -41,6 +41,8 @@ class VerificationCheckTestCase(IntegrationTestCase):
                 "channel": "sms",
                 "status": "approved",
                 "valid": false,
+                "amount": "$29.99",
+                "payee": "Acme",
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z"
             }