+++ twilio/rest/api/v2010/account/usage/record/monthly.py
@@ -207,6 +207,7 @@ class MonthlyInstance(InstanceResource):
     """  """
 
     class Category(object):
+        AGENT_CONFERENCE = "agent-conference"
         ANSWERING_MACHINE_DETECTION = "answering-machine-detection"
         AUTHY_AUTHENTICATIONS = "authy-authentications"
         AUTHY_CALLS_OUTBOUND = "authy-calls-outbound"