+++ twilio/rest/api/v2010/account/usage/trigger.py
@@ -348,6 +348,7 @@ class TriggerInstance(InstanceResource):
     """  """
 
     class UsageCategory(object):
+        AGENT_CONFERENCE = "agent-conference"
         ANSWERING_MACHINE_DETECTION = "answering-machine-detection"
         AUTHY_AUTHENTICATIONS = "authy-authentications"
         AUTHY_CALLS_OUTBOUND = "authy-calls-outbound"