+++ twilio/rest/taskrouter/v1/workspace/task_queue/task_queue_real_time_statistics.py
@@ -187,6 +187,7 @@ class TaskQueueRealTimeStatisticsInstance(InstanceResource):
             'account_sid': payload['account_sid'],
             'activity_statistics': payload['activity_statistics'],
             'longest_task_waiting_age': deserialize.integer(payload['longest_task_waiting_age']),
+            'longest_task_waiting_sid': payload['longest_task_waiting_sid'],
             'task_queue_sid': payload['task_queue_sid'],
             'tasks_by_priority': payload['tasks_by_priority'],
             'tasks_by_status': payload['tasks_by_status'],
@@ -241,6 +242,14 @@ class TaskQueueRealTimeStatisticsInstance(InstanceResource):
         :rtype: unicode
         """
         return self._properties['longest_task_waiting_age']
+
+    @property
+    def longest_task_waiting_sid(self):
+        """
+        :returns: The SID of the longest waiting Task
+        :rtype: unicode
+        """
+        return self._properties['longest_task_waiting_sid']
 
     @property
     def task_queue_sid(self):