+++ twilio/rest/taskrouter/v1/workspace/workspace_real_time_statistics.py
@@ -180,6 +180,7 @@ class WorkspaceRealTimeStatisticsInstance(InstanceResource):
             'account_sid': payload['account_sid'],
             'activity_statistics': payload['activity_statistics'],
             'longest_task_waiting_age': deserialize.integer(payload['longest_task_waiting_age']),
+            'longest_task_waiting_sid': payload['longest_task_waiting_sid'],
             'tasks_by_priority': payload['tasks_by_priority'],
             'tasks_by_status': payload['tasks_by_status'],
             'total_tasks': deserialize.integer(payload['total_tasks']),
@@ -231,6 +232,14 @@ class WorkspaceRealTimeStatisticsInstance(InstanceResource):
         :rtype: unicode
         """
         return self._properties['longest_task_waiting_age']
+
+    @property
+    def longest_task_waiting_sid(self):
+        """
+        :returns: The SID of the longest waiting Task
+        :rtype: unicode
+        """
+        return self._properties['longest_task_waiting_sid']
 
     @property
     def tasks_by_priority(self):